Projects_Assessments/Python/js-to-ts-transpilation/requirements.txt
flake8==3.9.0
pre-commit==2.9.0
openai==1.10.0
pytest==7.4.4
python_dotenv==1.0.1
transformers==4.37.1
tensorflow==2.14.0
sentencepiece==0.1.99
torch==2.1.2
accelerate==0.26.1
psutil==5.9.8